Add region editor add, remove, clear, rename
Add context menu to region editor items allowing new child regions to be added, or the current region to be cleared or removed. Removing the root region just clears it. Note that removing a region that was created by an ancestor model source will cause it - and similarly marked descendants - to be immediately recreated, however graphics and model sources on it and its descendants are lost.
Double click on region names to enter new names. Note regions cannot be renamed if they are the root, if the new name is an empty string, if an existing sibling region is using the new name, if the region was created by an ancestor model source (since it would have to be reloaded and it will reappear), or if Zinc denies the name change on the Zinc region.