Handle and report errors on loading Neon document
Logs errors in Logger and uses a message box to inform user of problem, referring to logger for more details.
If load failed, a blank document is created so not left in an unsafe state.
Loader checks version of Neon document and fails if file version is later than this Neon application.